day_02/pyproject.toml
[project]
name = "day-02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.11.0",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
]
